Android/NativeSensorListener: fix typos for two JNI functions
Apparently, nobody has been using a Nintento Nunchuk or a GliderLink
with XCSoar for a very long time!

+++ android/src/NativeSensorListener.java
@@ -81,7 +81,7 @@ final class NativeSensorListener implements SensorListener {
                                          double longitude, double latitude,
                                          double altitude,
                                          double gspeed, double vspeed,
-                                         double bearing);
+                                         int bearing);
 
   @Override
   public native void onTemperature(double temperature_kelvin);
